# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_QuerySandboxContainersResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuerySandboxContainersResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QuerySandboxContainersResponse]",
        "type": "object"
      },
...
      "CommonResponse_SandboxContainerSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxContainerSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[SandboxContainerSchema]",
        "type": "object"
      },
...
      "CreateSandboxContainerRequest": {
        "properties": {
          "egress_mode": {
            "$ref": "#/components/schemas/SandboxContainerEgressMode",
            "default": "direct"
          },
          "egress_proxy_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Egress Proxy Id"
          },
          "host_id": {
            "exclusiveMinimum": 0.0,
            "title": "Host Id",
            "type": "integer"
          },
          "image_id": {
            "exclusiveMinimum": 0.0,
            "title": "Image Id",
            "type": "integer"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Assign container owner user ID. Admin only; defaults to the creator.",
            "title": "Owner Id"
          },
          "port_mappings": {
            "items": {
              "$ref": "#/components/schemas/SandboxContainerPortMapping"
            },
            "maxItems": 32,
            "title": "Port Mappings",
            "type": "array"
          }
        },
        "required": [
          "host_id",
          "image_id"
        ],
        "title": "CreateSandboxContainerRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sandbox-containers": {
      "get": {
        "operationId": "query_sandbox_containers_route_api_sandbox_containers_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QuerySandboxContainersResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Sandbox Containers Route",
        "tags": [
          "sandbox-containers"
        ]
      },
      "post": {
        "operationId": "create_sandbox_container_route_api_sandbox_containers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxContainerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxContainerSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox image not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Sandbox Container Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/available": {
      "get": {
        "operationId": "query_available_sandbox_containers_route_api_sandbox_containers_available_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "work_project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Project Id"
            }
          },
          {
            "in": "query",
            "name": "include_non_running",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Non Running",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QuerySandboxContainersResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Available Sandbox Containers Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/create-options": {
      "get": {
        "operationId": "sandbox_container_create_options_route_api_sandbox_containers_create_options_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxContainerCreateOptionsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Sandbox Container Create Options Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/{id}": {
      "delete": {
        "operationId": "delete_sandbox_container_route_api_sandbox_containers__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteSandboxContainerResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox container not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Sandbox Container Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/{id}/egress": {
      "patch": {
        "operationId": "update_sandbox_container_egress_route_api_sandbox_containers__id__egress_patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSandboxContainerEgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxContainerSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox container not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Sandbox Container Egress Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/{id}/files": {
      "get": {
        "operationId": "list_container_files_route_api_sandbox_containers__id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "/",
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ListContainerFilesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox container not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "List Container Files Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
    "/api/sandbox-containers/{id}/files/copy": {
      "post": {
        "operationId": "copy_container_files_route_api_sandbox_containers__id__files_copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainerFileCopyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox container not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Copy Container Files Route",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
...
  }
}